Make shellbang universal
Bash might not be located in `/bin/bash` e.g. on Arch Linux and Debian
derivates it is in `/usr/bin/bash` and on other systems it might be
`/usr/local/bin/bash` or something completely different. Using
`/usr/bin/env` to locate the bash interpreter, the first occurrence of
`bash` in `$PATH` will be used